requirements.txt
aiofiles==0.8.0
anyio==3.5.0
apturl==0.5.2
asgiref==3.8.1
attrs==23.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
colorlog==6.8.2
command-not-found==0.3
coolercontrol-liqctld==1.3.0
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==42.0.5
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.8.0rc2
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2
django-colorfield==0.11.0
django-extensions==3.2.3
django-extra-fields==3.0.2
django-filter==24.2
django-modeltranslation==0.19.5
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
docopt==0.6.2
duplicity==0.8.21
fastapi==0.63.0
fasteners==0.14.1
flake8==6.0.0
flake8-isort==6.0.0
future==0.18.2
h11==0.13.0
hidapi==0.9.0.post3
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
iniconfig==2.0.0
IsDigit==0.0.10
isort==5.13.2
itsdangerous==2.1.0
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
liquidctl==1.13.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
MarkupSafe==2.1.5
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
packaging==24.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.3.0
pluggy==1.4.0
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py==1.11.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydantic==1.8.2
pyflakes==3.0.1
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==7.1.2
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.10.1
python3-openid==3.2.0
pytz==2024.1
pyusb==1.2.1.post1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==2.0.0
SecretStorage==3.3.1
setproctitle==1.2.2
simplejson==3.17.6
six==1.16.0
smbus==1.1.post2
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.5.3
sqlparse==0.4.4
starlette==0.18.0
systemd-python==234
tomli==2.0.1
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.1
usb-creator==0.3.7
uvicorn==0.15.0
vboxapi==1.0
wadllib==1.3.6
webcolors==1.13
wsproto==1.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
